requirements.txt
opencv-python-headless
easyocr==1.5.0
pyyaml
aiogram
python-dotenv